requirements/local.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2
django-widget-tweaks==1.4.12
idna==3.4
mysql-connector-python==8.0.31
mysqlclient==2.1.1
Pillow==9.2.0
protobuf==3.20.1
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
